vantage6-client/client/requirements.txt
requests
pyyaml
pyjwt